testing and other

+++ database/deployment/get_repo.go
@@ -15,7 +15,7 @@ import (
 )
 
 // GetDeploymentForRepo gets a deployment by repo ID and number from the database.
-func (e *engine) GetDeploymentForRepo(ctx context.Context, r *library.Repo, number int) (*library.Deployment, error) {
+func (e *engine) GetDeploymentForRepo(ctx context.Context, r *library.Repo, number int64) (*library.Deployment, error) {
 	e.logger.WithFields(logrus.Fields{
 		"deployment": number,
 		"org":        r.GetOrg(),
